データセット下処理を軽量化

+++ utils_common/prepare_dataset_finetune.py
@@ -10,11 +10,11 @@ import random
 DATA_MODE = "data_bt" # data_bt | data_parallel
 # CONSTRAINT_TYPE = "reference" # reference | random | source
 
-DATA_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/data_sources/{}/modified/'.format(DATA_MODE)
-OUT_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/datasets/{}/'.format(DATA_MODE)
+DATA_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/data_sources/{}/full/'.format(DATA_MODE)
+# DATA_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/data_sources/{}/'.format(DATA_MODE) # 実験のために一部のデータで処理を確認
+OUT_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/datasets/{}/full'.format(DATA_MODE) 
 # CONST_OUT_DIR = '/raid/ieda/trans_jaen_dataset/Dataset/datasets/{}/constraints/{}/'.format(DATA_MODE, CONSTRAINT_TYPE)
 CONST_OUT_DIR = os.path.join(OUT_DIR, 'constraints')
-
 
 def create_if_not_exist(p):
     if not os.path.exists(p):
@@ -28,15 +28,40 @@ assert os.path.exists(CONST_OUT_DIR)
 
 from utils import (
     ls, jpath,
-    SyllableCounterJA,
+    # SyllableCounterJA,
     BoundaryUtilEn,
-    RhymeUtil, # for Ja
+    # RhymeUtil, # for Ja
     PosSeg, # for Ja
     SyllableCounter,
     BoundaryUtil, # for Ja
     StressUtilEn,
 )
 from tqdm import tqdm
+import spacy
+
+nlp = spacy.load("ja_ginza")
+small_word = ["ャ","ュ","ョ","ァ","ィ","ゥ","ェ","ォ"]
+
+rhyme_6_dic = {
+    1:["ア","カ","サ","タ","ナ","ハ","マ","ヤ","ラ","ワ","ァ","ャ","ガ","ザ","ダ","バ","パ"], #ア行 "a"
+    2:["イ","キ","シ","チ","ニ","ヒ","ミ","リ","ィ","ギ","ジ","ヂ","ビ","ピ"], #イ行 "i"
+    3:["ウ","ク","ス","ツ","ヌ","フ","ム","ユ","ル","ゥ","ュ","グ","ズ","ヅ","ブ","プ"], #ウ行 "u"
+    4:["エ","ケ","セ","テ","ネ","ヘ","メ","レ","ェ","ゲ","ゼ","デ","ベ","ペ"], #エ行 "e"
+    5:["オ","コ","ソ","ト","ノ","ホ","モ","ヨ","ロ","ヲ","ォ","ョ","ゴ","ゾ","ド","ボ","ポ"], #オ行 "o"
+    6:["ン"], #ン行 "n"
+}
+
+# Get reverse dict of rhyme_6_dic
+t = {}
+for k in rhyme_6_dic:
+    for v in rhyme_6_dic[k]:
+        t[v] = k
+rhyme_6_dic_reverse = t
+
+special_alph = {"ッ", "ー"}
+
+
+error_log = []
 
 def _main():
     print("Start procedures")
@@ -46,36 +71,40 @@ def _main():
     print("Finish val procedures")
     test_procedures()
     print("Finish test procedures")
+    write_error_log()
 
 def train_procedures():
     train_source = load_save_text('train.source')
     train_target = load_save_text('train.target')
-    reference_rhylen(texts=train_target, mode='train') # create train.target
-    reference_boundary(texts=train_target, mode='train') # create train_boundary.target
-    reference_stress(texts=train_target, mode='train') # create train_stress.target
+    train_kana = translate2kana(train_target, 'train')
+    reference_rhylen(texts=train_kana, mode='train') # create train.target
+    # reference_boundary(texts=train_target, mode='train') # create train_boundary.target
+    # reference_stress(texts=train_target, mode='train') # create train_stress.target
     random_rhylen(texts=train_target, mode='train') # create /random/train.target
 
 def val_procedures():
     val_source = load_save_text('val.source')
     val_target = load_save_text('val.target')
-    reference_rhylen(texts=val_target, mode='val') # create /reference/val.target
-    reference_boundary(texts=val_target, mode='val') # create /reference/val_boundary.target
-    reference_stress(texts=val_target, mode='val') # create /reference/val_stress.target
+    val_kana = translate2kana(val_target, 'val')
+    reference_rhylen(texts=val_kana, mode='val') # create /reference/val.target
+    # reference_boundary(texts=val_target, mode='val') # create /reference/val_boundary.target
+    # reference_stress(texts=val_target, mode='val') # create /reference/val_stress.target
     random_rhylen(texts=val_target, mode='val') # create /random/val.target
     source_rhylen(texts=val_source, mode='val') # create /source/val.target
-    source_boundary(texts=val_source, mode='val') # create /source/val_boundary.target
-    source_stress(texts=val_source, mode='val') # create /source/val_stress.target
+    # source_boundary(texts=val_source, mode='val') # create /source/val_boundary.target
+    # source_stress(texts=val_source, mode='val') # create /source/val_stress.target
 
 def test_procedures():
     test_source = load_save_text('test.source')
     test_target = load_save_text('test.target')
-    reference_rhylen(texts=test_target, mode='test') # create /reference/test.target
-    reference_boundary(texts=test_target, mode='test') # create /reference/test_boundary.target
-    reference_stress(texts=test_target, mode='test') # create /reference/test_stress.target
+    test_kana = translate2kana(test_target, 'test')
+    reference_rhylen(texts=test_kana, mode='test') # create /reference/test.target
+    # reference_boundary(texts=test_target, mode='test') # create /reference/test_boundary.target
+    # reference_stress(texts=test_target, mode='test') # create /reference/test_stress.target
     random_rhylen(texts=test_target, mode='test') # create /random/test.target
     source_rhylen(texts=test_source, mode='test') # create /source/test.target
-    source_boundary(texts=test_source, mode='test') # create /source/test_boundary.target
-    source_stress(texts=test_source, mode='test') # create /source/test_stress.target
+    # source_boundary(texts=test_source, mode='test') # create /source/test_boundary.target
+    # source_stress(texts=test_source, mode='test') # create /source/test_stress.target
 
 def load_save_text(file_name): # load texts from dataset from "data_source", save to "datasets"
     # load texts from dataset
@@ -86,18 +115,73 @@ def load_save_text(file_name): # load texts from dataset from "data_source", sav
 
     # save texts to directory "datasets"
     out_path = jpath(OUT_DIR, file_name)
+    if os.path.exists(out_path):
+        return texts
     with open(out_path, 'w', encoding='utf8') as f:
         f.writelines(texts)
     return texts
 
+def translate2kana(texts,mode):
+    out_path = jpath(OUT_DIR, mode + '.kana')
+    if os.path.exists(out_path):
+        with open(out_path, 'r', encoding='utf8') as f:
+            return f.readlines()
+    # create kana lines
+    kana_lines = []
+    for sentence in tqdm(texts):
+        try:
+            doc = nlp(sentence)
+            line = ""
+            for token in doc:
+                if token.pos_ in ['PUNCT', 'SYM', 'SPACE', 'X']:
+                    continue
+                # print(token.morph.get("Reading"))
+                if token.morph.get("Reading") == []:
+                    continue
+                line += token.morph.get("Reading")[0]
+            kana_lines.append(line + "\n")
+            # print(line)
+        except Exception as e:
+            print(e)
+            exit()
+    with open(out_path, "w") as fo:
+        fo.writelines(kana_lines)
+    return kana_lines
+
+def calc_syllables(kana_line):
+    length = 0
+    for ch in list(kana_line.strip()):
+        if not ch in small_word:
+            length += 1
+    return length
+
+def calc_rhyme_type(kana_line):
+    last_ch = kana_line.strip()[-1]
+    if last_ch in special_alph:
+        last_ch = kana_line.strip()[-2]
+    return rhyme_6_dic_reverse[last_ch]
+
 def reference_rhylen(texts, mode):
     out_path = jpath(CONST_OUT_DIR, "reference", mode + '.target')
     # Create rhyme and length constraints
-    rhyme_util = RhymeUtil()
+    # rhyme_util = RhymeUtil()
     consts_rhy_len = []
-    for line in tqdm(texts):
-        length = SyllableCounterJA.count_syllable_sentence(line.strip())
-        rhyme = rhyme_util.get_rhyme_type_of_line(line.strip())
+    for i,line in enumerate(tqdm(texts)):
+        try:
+            # length = SyllableCounterJA.count_syllable_sentence(line.strip())
+            length = calc_syllables(line)
+            if length > 20:
+                error_log.append("Error in line {} of {}: {}".format(i, mode, line))
+                length = 20
+        except Exception as e:
+            length = random.randint(3, 20)
+            error_log.append("Error in line {} of {}: {}".format(i, mode, line))
+        try:
+            # rhyme = rhyme_util.get_rhyme_type_of_line(line.strip())
+            rhyme = calc_rhyme_type(line)
+        except Exception as e:
+            rhyme = random.randint(1, 6)
+            error_log.append("Error in line {} of {}: {}".format(i, mode, line))
         consts_rhy_len.append('{}\t{}'.format(length, rhyme))
     with open(out_path, 'w', encoding='utf-8') as f:
         f.writelines([i + '\n' for i in consts_rhy_len])
@@ -143,10 +227,18 @@ def source_rhylen(texts,mode):
     # Create rhyme and length constraints
     # rhyme is randomly generated (English sentence cannot decide Japanese rhyme type)
     out_path = jpath(CONST_OUT_DIR, "source", mode + '.target')
-    syllable_util = SyllableCounterJA()
+    syllable_util = SyllableCounter()
     consts_rhy_len = []
-    for line in tqdm(texts):
-        length = syllable_util.count_syllable_sentence(line.strip())
+    for i,line in enumerate(tqdm(texts)):
+        try:
+            length = syllable_util.count_syllable_sentence(line.strip())
+            # length = calc_syllables(line)
+            # if length > 20:
+            #     error_log.append("Error in line {} of {}: {}".format(i, mode, line))
+            #     length = 20
+        except Exception as e:
+            length = random.randint(3, 20)
+            error_log.append("Error in line {} of {}: {}".format(i, mode, line))
         rhyme = random.randint(1, 6)
         consts_rhy_len.append('{}\t{}'.format(length, rhyme))
     with open(out_path, 'w', encoding='utf-8') as f:
@@ -176,6 +268,13 @@ def source_stress(texts,mode):
         consts_str.append(stress)
     with open(out_path, 'w', encoding='utf-8') as f:
         f.writelines([i + '\n' for i in consts_str])
+
+def write_error_log():
+    if not error_log:
+        return
+    with open(os.path.join(OUT_DIR, 'error_log.txt'), 'w', encoding='utf8') as f:
+        f.writelines(error_log)
+    return
 
 if __name__ == '__main__':
     _main()
